+++ package.json
@@ -10,10 +10,12 @@
   },
   "dependencies": {
     "clsx": "^2.1.1",
+    "date-fns": "^4.1.0",
     "lodash.debounce": "^4.0.8",
     "next": "15.3.1",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "svgr": "^1.10.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
